src/auth/requirements.txt
astroid==3.3.3
blinker==1.8.2
click==8.1.7
colorama==0.4.6
dill==0.3.8
Flask==3.0.3
Flask-MySQLdb==2.0.0
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
MarkupSafe==2.1.5
mccabe==0.7.0
mysqlclient==2.2.4
parso==0.8.4
platformdirs==4.3.6
PyJWT==2.9.0
pylint==3.3.0
tomlkit==0.13.2
Werkzeug==3.0.4
